refactor(vm_extensions): move _create_and_verify_extension_run into base
Deduplicate the three identical/near-identical _create_and_verify_extension_run copies (CustomScript, RunCommand v1/v2) into a single VmExtensionTestBase method. Add a SUPPORTS_DELETE class flag (False for CRP-managed RunCommand v2) to gate the pre-delete step, and drop the now-unused assert_that, execute_command, and Optional imports from the suites.

[AI Generated]
